bonus/demo_env_tk_alchemy/requirements.txt
python-dotenv>=1.0,<2.0
SQLAlchemy>=2.0,<3.0